Documentation for debugger mode
This sets up the structure for the documentation for the new debug mode;
it's not 100% complete yet. I also added a bunch of TODOs There are a
bunch of TODOs sprinkled around for stuff I noticed that behaved
unexpected/buggy.

I added it as a new section instead of putting it in the regular
commands/etc. sections. This seemed to make the most sense to me
personally: this way there is an obvious location for a "getting
started" section, and we can keep the docs for :GoDebugStart as a short
and to the point reference doc.It also avoids having to add "This
command/mapping is only available after using :GoDebug" on all the
commands.

I can move stuff around if you want.