Add :GoDebugWindow command (WIP)
My aim was to make it possible to allow users to dynamically create the
debug windowswith :GoDebugWindow. This also adds commands to get the log
output etc. from a command (:GoDebugMessages).

Very work-in-progress! I thought this would be a simple change but turns
out it's not. Not sure if I'll finish this